Rack::File#serving now returns Rack::File::Iterator
Background:
==========

In the `body` Sinatra helper, we DON'T delete the 'Content-Length'
header when:

1. the request is a HEAD request OR
2. If the "value" is a `Rack::File` type OR
3. If the "value" is a `Stream`

where "value" is the argument passed to the `body` helper method. (Refer
https://github.com/sinatra/sinatra/issues/739 and
https://github.com/sinatra/sinatra/pull/770)

Post Rack 2.0, the `Rack::File#serving` method now returns a response
array with the second argument of type `Rack::File::Iterator`. This
change breaks this conditional and deletes the `Content-Length` header.
This changes fixes this problem.